Fjerner funksjonalitet for query-basert batch migrering til layers
